% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbundle-print.R
\name{field_names}
\alias{field_names}
\title{Names of VarBundle}
\usage{
field_names(x)
}
\arguments{
\item{x}{(VarBundle)}
}
\value{
(character vector)
}
\description{
Returns field names of VarBundle Object
}
