% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbundle-print.R
\name{print.VarBundle}
\alias{print.VarBundle}
\title{Print VarBundle}
\usage{
\method{print}{VarBundle}(x, ...)
}
\arguments{
\item{x}{(VarBundle)}

\item{...}{(additional args)}
}
\description{
Prints a VarBundle object.
}
